Update for Lab12.c
