+++ PSC.sln
@@ -25,6 +25,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Lab10", "Lab10\Lab10.vcxpro
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Lab11", "Lab11\Lab11.vcxproj", "{BCBD6024-49D2-429B-969A-C53A87A94D70}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Lab12", "Lab12\Lab12.vcxproj", "{34DDAA9D-1391-43B3-A457-2250737F4F1D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -121,6 +123,14 @@ Global
 		{BCBD6024-49D2-429B-969A-C53A87A94D70}.Release|x64.Build.0 = Release|x64
 		{BCBD6024-49D2-429B-969A-C53A87A94D70}.Release|x86.ActiveCfg = Release|Win32
 		{BCBD6024-49D2-429B-969A-C53A87A94D70}.Release|x86.Build.0 = Release|Win32
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Debug|x64.ActiveCfg = Debug|x64
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Debug|x64.Build.0 = Debug|x64
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Debug|x86.ActiveCfg = Debug|Win32
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Debug|x86.Build.0 = Debug|Win32
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Release|x64.ActiveCfg = Release|x64
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Release|x64.Build.0 = Release|x64
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Release|x86.ActiveCfg = Release|Win32
+		{34DDAA9D-1391-43B3-A457-2250737F4F1D}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
